+++ ADAUGDB/views/open_elective_rollList.py
@@ -31,7 +31,7 @@ def open_elective_rollList(request):
                 asem = int(strs[4])
                 byear = rom2int[strs[1]]
                 bsem = rom2int[strs[2]]
-                regulation = int(strs[5])
+                regulation = float(strs[5])
                 mode = strs[6]
                 file = form.cleaned_data.get('file')
                 BTOpenElectiveRollLists.objects.filter(RegEventId__AYear=ayear,RegEventId__ASem=asem,RegEventId__BYear=byear,RegEventId__BSem=bsem,RegEventId__Regulation=regulation,RegEventId__Mode=mode,subject_id=subid).delete()
@@ -107,7 +107,7 @@ def OERollList_Status(request):
                 asem = int(strs[4])
                 byear = rom2int[strs[1]]
                 bsem = rom2int[strs[2]]
-                regulation = int(strs[5])
+                regulation = float(strs[5])
                 mode = strs[6]
                 subid = form.cleaned_data.get('sub')
                 rows = BTOpenElectiveRollLists.objects.filter(subject_id=subid,RegEventId__AYear=ayear,RegEventId__ASem=asem,RegEventId__BYear=byear,RegEventId__BSem=bsem,RegEventId__Regulation=regulation,RegEventId__Mode=mode).order_by('student__id')
